Refactored subcommands for the cms command into their own subdirectory Removed deprecated dbgettext documentation Added `cms moderator on` command (+ tests)
